Refactor routine ID type from Long to Int and update related logic; add notification scheduling and cancellation use cases

+++ app/src/main/java/fr/equipe8/projetinfomobile/data/routines/RoutineRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 class RoutineRepository (private val dao: RoutineDao) {
     //Fast external functions
     fun getAllRoutines(): Flow<List<Routine>> = dao.getAllRoutines()
-    fun getRoutineById(routineId: Long): Routine? = dao.getRoutineById(routineId)
+    fun getRoutineById(routineId: Int): Routine? = dao.getRoutineById(routineId)
     suspend fun upsertRoutine(routine: Routine) = dao.insert(routine)
     suspend fun removeRoutine(routine: Routine) = dao.delete(routine)
