+++ app/src/main/java/fr/equipe8/projetinfomobile/ui/routinescreen/RoutinesListEvent.kt
@@ -1,8 +1,8 @@
 package fr.equipe8.projetinfomobile.ui.routinescreen
 
 sealed class RoutinesListEvent {
-    data class RoutineClicked(val id : Long) : RoutinesListEvent()
-    data class ActiveRoutineClicked(val id : Long) : RoutinesListEvent()
+    data class RoutineClicked(val id : Int) : RoutinesListEvent()
+    data class ActiveRoutineClicked(val id : Int) : RoutinesListEvent()
     data object NewRoutine : RoutinesListEvent()
     data object  DebugMessage : RoutinesListEvent()
 }
