+++ app/src/main/java/fr/equipe8/projetinfomobile/ui/routinescreen/RoutinesListUIEvent.kt
@@ -2,5 +2,5 @@ package fr.equipe8.projetinfomobile.ui.routinescreen
 
 sealed interface RoutinesListUiEvent {
     data class ShowMessage(val message: String) : RoutinesListUiEvent
-    data class Navigate(val id : Long) : RoutinesListUiEvent
+    data class Navigate(val id : Int) : RoutinesListUiEvent
 }
