+++ app/src/main/java/fr/equipe8/projetinfomobile/usecases/GetRoutineByIdUseCase.kt
@@ -4,7 +4,7 @@ import fr.equipe8.projetinfomobile.data.routines.Routine
 import fr.equipe8.projetinfomobile.data.routines.RoutineRepository
 
 class GetRoutineByIdUseCase(private val repository: RoutineRepository) {
-    operator fun invoke(id : Long): Routine? {
+    operator fun invoke(id : Int): Routine? {
         return repository.getRoutineById(id)
     }
 }
